add style

+++ src/Components/SinglePost/Comment/Comment.tsx
@@ -1,7 +1,8 @@
 import React, { useState } from 'react'
 import { useAppSelector } from '../../../hook/hook'
 import { commentApi } from '../../../API/api'
-
+import style from './Comment.module.scss'
+import { DeleteOutlined, EditOutlined } from '@ant-design/icons'
 type Props = {
    id: number
    body: string
@@ -19,11 +20,13 @@ const Comment = ({ body, author, created, id, getSinglePost, image }: Props) => 
    const [edit, setEdit] = useState(false)
    const [commentBody, setCommentBody] = useState(body)
    const deletePost = (id: number) => {
-      commentApi.deleteComment(id).then(
-         () => {
-            getSinglePost()
-         }
-      )
+      if (window.confirm('delete comment?')) {
+         commentApi.deleteComment(id).then(
+            () => {
+               getSinglePost()
+            }
+         )
+      }
    }
    const editPost = (id: number) => {
       if (commentBody !== body) {
@@ -39,15 +42,23 @@ const Comment = ({ body, author, created, id, getSinglePost, image }: Props) => 
 
    let user = useAppSelector((state) => state.userReducer.name)
    return (
-      <div>
-         <p>{author} {created.slice(0, 10)}</p>
-         <div>
+      <div className={style.comment}>
+         <div className={style.comment__info}>
+            <div className={style.comment__author}>
+               <p>{author}</p>
+               <p>{created.slice(0, 10)}</p>
+            </div>
+            <div className={style.comment__edit}>
+               {user === author && <EditOutlined onClick={() => setEdit(true)} />}
+               {user === author && <DeleteOutlined onClick={() => deletePost(id)} />}
+            </div>
+
+         </div>
+         <div className={style.comment__body}>
             {!edit && <p> {body}</p>}
-            {edit && <p> <input type="text" value={commentBody} autoFocus={true} onChange={(e) => { setCommentBody(e.target.value) }} onBlur={() => editPost(id)} /> </p>}
+            {edit && <p> <textarea value={commentBody} autoFocus={true} onChange={(e) => { setCommentBody(e.target.value) }} onBlur={() => editPost(id)} /> </p>}
          </div>
          <p>{image && <img src={image} alt="" />}</p>
-         <p>{user === author && <button onClick={() => deletePost(id)}>delete</button>}</p>
-         <p>{user === author && <button onClick={() => setEdit(true)}>edit</button>}</p>
       </div>
    )
 }
